backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pillow==10.2.0
pytesseract==0.3.10
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.1
pydantic-settings==2.2.1
slowapi==0.1.9
boto3==1.34.0
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0